backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
requests==2.32.2
python-dotenv==1.0.1
